requirements.txt
aggdraw==1.3.12
anyio==2.2.0
appnope==0.1.2
argon2-cffi==20.1.0
# astroid @ file:///opt/concourse/worker/volumes/live/a8a8b2be-963f-4e30-7c46-4aecaf981ab4/volume/astroid_1613500845571/work
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
contextvars==2.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
deprecation==2.1.0
entrypoints==0.3
future==0.18.2
gitdb==4.0.7
GitPython==3.1.14
idna==2.10
immutables==0.15
importlib-metadata==3.5.0
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
# isort @ file:///tmp/build/80754af9/isort_1612243941122/work
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyter-packaging==0.9.2
jupyter-server==1.6.4
jupyter-server-mathjax==0.2.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
# lazy-object-proxy @ file:///opt/concourse/worker/volumes/live/4c22169c-6e74-4168-5f7b-2b5799eeffaa/volume/lazy-object-proxy_1606850079154/work
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.2
nbconvert==6.0.7
nbdime==3.0.0
nbformat==5.1.2
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.2.0
numpy==1.19.5
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.0
# pylint @ file:///opt/concourse/worker/volumes/live/8f9e8f73-8ae9-4550-7ce6-1e4ba902c47d/volume/pylint_1614971221136/work
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
regex==2021.4.4
requests==2.25.1
# rope @ file:///tmp/build/80754af9/rope_1602264064449/work
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.5.0
Shapely==1.7.1
# six @ file:///opt/concourse/worker/volumes/live/5839b2d2-7415-4ff0-4c61-c4cddf798f7c/volume/six_1605205320479/work
sklearn==0.0
smmap==4.0.0
sniffio==1.2.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
# toml @ file:///tmp/build/80754af9/toml_1592853716807/work
tomlkit==0.7.0
torch==1.2.0
torchsummary==1.5.1
torchvision==0.4.0
tornado==6.1
tqdm==4.57.0
traitlets==4.3.3
# typed-ast @ file:///opt/concourse/worker/volumes/live/85c87157-84ad-443c-5815-d7450a5595c1/volume/typed-ast_1610484564230/work
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yapf==0.31.0
zipp==3.4.0
